requirements.txt
absl-py==2.1.0
attrs==23.2.0
cffi==1.16.0
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
dlib @ file:///C:/Users/Administrator/Desktop/Robo-dog/Python_robot_dog-tower/venv/Include/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
face-recognition==1.3.0
face_recognition_models==0.3.0
flatbuffers==24.3.25
fonttools==4.50.0
jax==0.4.25
kiwisolver==1.4.5
matplotlib==3.8.3
mediapipe==0.10.11
ml-dtypes==0.3.2
numpy==1.24.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
Pillow==9.4.0
platformdirs==4.2.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
scipy==1.12.0
six==1.16.0
sounddevice==0.4.6
virtualenv==20.25.1
